Cria a base de casos para serem simulados posteriormente no cálculo do payback.

Cria a base de casos para serem simulados posteriormente no cálculo do payback.

epe4md_casos_payback(
  ano_base,
  ano_max_resultado = 2060,
  inflacao = 0.0375,
  ano_troca_inversor = 11,
  fator_custo_inversor = 0.15,
  dir_dados_premissas = NA_character_
)

Arguments

ano_base

numeric. Ano base da projeção. Define o ano em que a função irá buscar a base de dados. Último ano completo realizado.

ano_max_resultado

numeric. Ano final para apresentação dos resultados. Máximo igual a 2060. Default igual a 2060.

inflacao

numeric. Taxa de inflacão considerada no reajuste das tarifas e para calcular o retorno real de projetos.

ano_troca_inversor

numeric. Ano, a partir do ano de instalação, em que é realizada a troca do inversor fotovoltaico. Default igual a 11.

fator_custo_inversor

numeric. Custo do inversor para a substituição em percentual do CAPEX total do sistema de geração. Default igual a 0.15.

dir_dados_premissas

Diretório onde se encontram as premissas. Se esse parâmetro não for passado, a função usa os dados default que são instalados com o pacote. É importante que os nomes dos arquivos sejam os mesmos da pasta default.

Value

data.frame. Casos para serem simulados posteriormente no cálculo do payback.