Roda um fluxo de caixa para cada caso e retorna métricas financeiras.

Roda um fluxo de caixa para cada caso e retorna métricas financeiras.

epe4md_payback(
  casos_payback,
  premissas_reg,
  ano_base,
  altera_sistemas_existentes = TRUE,
  ano_decisao_alteracao = 2023,
  ano_final_alteracao = 2060,
  inflacao = 0.0375,
  taxa_desconto_nominal = 0.13,
  custo_reforco_rede = 200,
  ano_troca_inversor = 11,
  ano_recapex_bat = 11,
  degradacao_bateria_mil_ciclos = 0.1,
  pagamento_disponibilidade = 0.3,
  disponibilidade_kwh_mes = 100,
  desconto_capex_local = 0,
  anos_desconto = 0,
  tarifa_bonus = 0,
  ano_inicio_bonus = 2099,
  custo_deficit = 0,
  bateria_eficiencia = 0.9,
  dir_dados_premissas = NA_character_
)

Arguments

casos_payback

data.frame. Base gerada pela função epe4md_casos_payback

premissas_reg

data.frame. Input de premissas regulatórias para serem consideradas nos cálculos.

ano_base

numeric. Ano base da projeção. Define o ano em que a função irá buscar a base de dados. Último ano completo realizado.

altera_sistemas_existentes

logic. TRUE se alterações regulatórias afetam investimentos realizados em anos anteriores à revisão da regulação. Default igual a FALSE.

ano_decisao_alteracao

numeric. Ano em que são definidas novas regras e se tornam de conhecimento público. Esse parâmetro só tem efeito caso o anterior seja igual a TRUE. Default igual a 2023.

ano_final_alteracao

numeric. Até qual ano as alterações regulatórias previstas afetam investimentos presentes. Só válido quando altera_sistemas_existentes igual a TRUE. Default igual a 2060.

inflacao

mumeric. Taxa anual de inflacao considerada no reajuste das tarifas e para calcular o retorno real de projetos. Default igual a 0.0375.

taxa_desconto_nominal

numeric. Taxa de desconto nominal considerada nos cálculos de payback descontado. Default igual a 0.13.

custo_reforco_rede

numeric. Custo em R$/kW aplicado a projetos de geracao remota em Alta Tensão. Representa um custo pago pelo empreendedor para reforços na rede. Default igual a 200.

ano_troca_inversor

numeric. Ano, a partir do ano de instalação, em que é realizada a troca do inversor fotovoltaico. Default igual a 11.

ano_recapex_bat

numeric. Ano em que será feito um investimento adicional em baterias para compensar a degradação. Default igual a 11.

degradacao_bateria_mil_ciclos

numeric. Degradação linear da bateria, em percentual a cada 1000 ciclos. Default igual a 10%.

disponibilidade_kwh_mes

numeric. Consumo de disponbilidade do consumidor em kWh/mês. Default igual a 100, equivalente a um consumidor trifásico. Tem efeito somente até o ano de 2022.

desconto_capex_local

numeric. Percentual de desconto a ser aplicado no CAPEX de sistemas de geração local(ex: 0.1) para simulação de incentivos. Default igual a 0.

anos_desconto

vector. Anos em que há a incidência do desconto no CAPEX.Ex: c(2024, 2025). Default igual a 0.

tarifa_bonus

integer. Tarifa que representa benefícios adicionais da MMGD. Valor em R$/kWh. A tarifa é multiplicada pela energia injetada na rede para formar uma receita adicional ao empreendimento. Default igual a 0.

ano_inicio_bonus

integer. Ano em que o bônus começa a ser incorporado na receita.

custo_deficit

numeric. Custo atribuído pelo consumidor à interrupção de energia. Em R$/MWh. Default = 0.

bateria_eficiencia

numeric. Eficiência das baterias. Default igual a 0.90.

dir_dados_premissas

Diretório onde se encontram as premissas. Se esse parâmetro não for passado, a função usa os dados default que são instalados com o pacote. É importante que os nomes dos arquivos sejam os mesmos da pasta default.

pagamento_disponibilidade.

numeric. Percentual de meses em que o consumidor residencial paga custo de disponbilidade em função da variabilidade da geração FV. Default igual a 0.3. Tem efeito somente até o ano de 2022.

Value

data.frame. Métricas financeiras para cada caso.