Roda o modelo 4MD

O resultado do modelo 4MD são projeções de capacidade instalada, número de adotantes e geração de energia em base mensal.

epe4md_calcula(
  premissas_reg,
  ano_base,
  ano_max_resultado = 2050,
  altera_sistemas_existentes = FALSE,
  ano_decisao_alteracao = 2023,
  inflacao = 0.0375,
  taxa_desconto_nominal = 0.13,
  custo_reforco_rede = 200,
  ano_troca_inversor = 11,
  pagamento_disponibilidade = 0.3,
  disponibilidade_kwh_mes = 100,
  filtro_renda_domicilio = "maior_3sm",
  desconto_capex_local = 0,
  anos_desconto = 0,
  tx_cresc_grupo_a = 0.016,
  spb = 0.3,
  p_max = 0.01,
  q_max = 1,
  filtro_comercial = NA_real_,
  ajuste_ano_corrente = FALSE,
  ultimo_mes_ajuste = NA_integer_,
  metodo_ajuste = NA_character_,
  dir_dados_premissas = NA_character_
)

Arguments

premissas_reg

data.frame. Input de premissas regulatórias para serem consideradas nos cálculos. O dataframe deve ter as seguintes colunas

ano, numérico

alternativa, numérico. Uma das seguintes opções:

0: Consumidor compensa todas as componentes tarifárias;

1: Paga TUSD Distribuição;

2: Anterior + TUSD Transmissão.

3: Anterior + TUSD Encargos.

4: Anterior + TUSD Perdas.

5: Anterior + TE Encargos. Ou seja, compensa somente a TE Energia.

p_transicao, numérico. Parcela do custo da alternativa escolhida no parâmetro alternativa a ser pago pelo consumidor

binomia e, binário. Define se há cobrança de uma tarifa binômia na baixa tensão, em que as componentes TUSD Distribuição e TUSD Transmissão passariam a ser cobradas de forma fixa, não sendo passíveis de compensação

demanda_g, binário. Define se há cobrança de TUSDg para a demanda de consumidores do grupo A. Caso seja FALSE, é considerada a cobrança da TUSD consumo.

Um arquivo excel instalado com este pacote, acessível via system.file("dados_premissas/2021/premissas_reg.xlsx", package = "epe4md"), contém um exemplo de premissas de entrada.

ano_base

numeric. Ano base da projeção. Define o ano em que a função irá buscar a base de dados. Último ano completo realizado.

ano_max_resultado

numeric. Ano final para apresentação dos resultados. Máximo igual a 2050. Default igual a 2050.

altera_sistemas_existentes

logic. TRUE se alterações regulatórias afetam investimentos realizados em anos anteriores à revisão da regulação. Default igual a FALSE.

ano_decisao_alteracao

numeric. Ano em que são definidas novas regras e se tornam de conhecimento público. Esse parâmetro só tem efeito caso o anterior seja igual a TRUE. Default igual a 2023.

inflacao

mumeric. Taxa anual de inflacao considerada no reajuste das tarifas e para calcular o retorno real de projetos. Default igual a 0.0375.

taxa_desconto_nominal

numeric. Taxa de desconto nominal considerada nos cálculos de payback descontado. Default igual a 0.13.

custo_reforco_rede

numeric. Custo em R$/kW aplicado a projetos de geracao remota em Alta Tensão. Representa um custo pago pelo empreendedor para reforços na rede. Default igual a 200.

ano_troca_inversor

numeric. Ano, a partir do ano de instalação, em que é realizada a troca do inversor fotovoltaico. Default igual a 11.

disponibilidade_kwh_mes

numeric. Consumo de disponbilidade do consumidor em kWh/mês. Default igual a 100, equivalente a um consumidor trifásico.

filtro_renda_domicilio

string. Define o filtro aplicado a consumidores residenciais, de acordo com a renda mensal do responsável, em salários mínimos. Permite: "total", "maior_1sm", maior_2sm", "maior_3sm" ou "maior_5sm". Default igual a "maior_3sm".

desconto_capex_local

numeric. Percentual de desconto a ser aplicado no CAPEX de sistemas de geração local(ex: 0.1) para simulação de incentivos. Default igual a 0.

anos_desconto

vector. Anos em que há a incidência do desconto no CAPEX. Default igual a 0.

tx_cresc_grupo_a

numeric. Taxa de crescimento anual dos consumuidores cativos do Grupo A. Default igual a 0.016 representa crescimento entre 2006 e 2019.

spb

numeric. Fator de Sensibilidade ao Payback (SPB). Default igual a 0.3.

p_max

numeric. Fator de inovação (p) máximo. Default igual a 0.01.

q_max

numeric. Fator de imitação (q) máximo. DEfault igual a 1.

filtro_comercial

numeric. Fator percentual para definir o nicho do segmento comercial. Default é calculado pelo modelo com base no nicho residencial.

ajuste_ano_corrente

logic. Se TRUE indica que a projeção deverá incorporar o histórico mensal recente, verificado em parte do primeiro ano após o ano base. Default igual a FALSE. O arquivo base_mmgd.xlsx deve incorporar esse histórico.

ultimo_mes_ajuste

numeric. Último mês com dados completos na base_ano_corrente. Default igual a NA. Só tem efeito caso ajuste_ano_corrente seja igual a TRUE.

metodo_ajuste

string. Se igual a "extrapola" o modelo irá extrapolar a potência e o número de adotantes até o final do ano base + 1 com base no verificado até o ultimo_mes_ajuste. Default igual a NA. Só tem efeito caso ajuste_ano_corrente seja igual a TRUE.

dir_dados_premissas

Diretório onde se encontram as premissas. Se esse parâmetro não for passado, a função usa os dados default que são instalados com o pacote. É importante que os nomes dos arquivos sejam os mesmos da pasta default.

pagamento_disponibilidade.

numeric. Percentual de meses em que o consumidor residencial paga custo de disponbilidade em função da variabilidade da geração FV. Default igual a 0.3.

Value

data.frame com os resultados da projeção de capacidade instalada de micro e minigeração distribuída, número de adotantes e geração mensal de energia.